fix(kv): correct three overstated claims, extract the shared cut driver, clear on zero at b > 1 (#382)
Review follow-up. Mostly claims that said more than the code proved.

**The rotor/iso clamp asymmetry had a false justification.** The comment said
clamping there "would throw the ring tail away". It could not: a ring-only tail
spans `[blocks_coverage, shape[2])`, strictly below `shape[2]`, and the clamp is
`min(n, shape[2])`, so it only touches `n > shape[2]`. The real reason the
asymmetry is safe is that `synced_rotor_v_blocks` / `synced_iso_v_blocks` size
their ring readback from `shape[2]` and return `Err` when the ring cannot cover
it — those stores already abort loudly and the clamp buys them nothing.
Restated in `storage/mod.rs` and `docs/KV_QUANT.md`.

**`KvStorage::truncate_to` denied a divergence it creates.** The comment claimed
every store clamps internally; the rotor/iso stores deliberately do not. So for
`n > shape[2]` the mixed arms leave the two axes of one codec at different
lengths — `IsoV3/4` and `RotorV3/4` (affine K clamps, codec V does not) and
`RotorKAsym3/4` (the other way round). It surfaces on spill, where the layer
geometry comes from the K shape while the V payload is written raw. Named in the
comment instead of denied.

**The dispatch test claimed coverage it does not have.** Five arms are
exercised, which covers all six store types but not the enum: `K8V8`,
`K8VTurbo2`, both TCQ variants, the K axis of `IsoV3/4` and `RotorV3/4`, and the
V axis of `RotorKAsym3/4` are unpinned because nothing in the workspace drives
`truncate_to` on them. `reset` pins only its `K8V4` arm. Both the test doc and
`docs/KV_QUANT.md` now say so and name the gaps.

**`QuantK::retain_cpu_prefix` refused to clear at `b > 1`.** Since `reset` routes
every store through `truncate_to(0)`, a `b > 1` reset cleared the `QuantV` blocks
(the planner drops all blocks regardless of `b`) while leaving `QuantK::codes`
intact under `shape[2] == 0` — permanently un-decodable. A clear has no prefix to
get wrong, so `n <= 0` is now handled before both refusals.

**Five byte-identical cut bodies extracted.** `BlockRows` gains `rows()`, and
`storage::truncate_block_store` owns the clamp → plan → apply → `shape[2]`
sequence for the five block stores. All thirteen `truncate_plan` call sites now
ask for the row count the same way, so the two row-count conventions are not
re-derived per codec — that drift is what let twelve arms keep a bare
`shape[2] = n`.

Also:
- `probe_v_dequant` returns `Option<Result<..>>`. Collapsing a refusing dequant
  into `None` reported "no V buffer" for the exact failure mode this work
  introduces, and the mutation checks depend on reading it. Its `pub`
  justification was also wrong — `wildcard_enum_match_arm` does not fire on a
  `let … else`, as this branch's own dispatch test shows — and is replaced with
  the honest one: one dispatch instead of four.
- The hydrate round trip now asserts the K axis too. `QuantK::retain_cpu_prefix`
  is the only element-granular cut and the only one carrying refusals, and it
  had no post-hydrate assertion at all.
- Its anti-vacuity comment pointed at `decode_fp16_kv().is_none()`, a weaker
  predicate than the `decode_fp16_k.is_some()` the codec early-return keys on.
  The load is actually carried by the decode-length assertion; the comments now
  say which is which.
- New `hydrate_truncate_tests.rs` sibling (child of `hydrate_tests` so it reuses
  those fixtures) rather than growing a 1361-line file.
- Two `#[allow]` blocks whose `reason=` described code that was not there,
  deleted; clippy stays clean.
