+++ crates/rmlx-kv-quant/src/storage/quant_iso_v4.rs
@@ -169,7 +169,11 @@ impl QuantIsoV4 {
     /// an SSD spill can rebuild it via [`synced_iso_v_blocks`].
     pub fn truncate_to(&mut self, n: i32) {
         let n = n.max(0);
-        let plan = super::truncate_plan(self.blocks.iter().map(|blk| blk.n_tokens), &self.shape, n);
+        let plan = super::truncate_plan(
+            self.blocks.iter().map(super::BlockRows::rows),
+            &self.shape,
+            n,
+        );
         super::apply_truncate_plan(&mut self.blocks, &plan);
         // NB: no `self.gpu.clear()` — the ring is the source of truth for a
         // ring-only decode tail; see [`super::QuantIsoV3::truncate_to`].
