+++ crates/rmlx-kv-quant/src/storage/quant_rotor_v3.rs
@@ -67,6 +67,10 @@ impl RotorBlocks {
 }
 
 impl super::BlockRows for RotorBlocks {
+    fn rows(&self) -> usize {
+        self.n_tokens
+    }
+
     /// The exhaustive destructure is the drift guard: a new payload field
     /// cannot be added without this failing to compile, which is what stops a
     /// buffer from surviving a mid-block truncation at its full length.
@@ -303,7 +307,11 @@ impl QuantRotorV3 {
     /// path.
     pub fn truncate_to(&mut self, n: i32) {
         let n = n.max(0);
-        let plan = super::truncate_plan(self.blocks.iter().map(|blk| blk.n_tokens), &self.shape, n);
+        let plan = super::truncate_plan(
+            self.blocks.iter().map(super::BlockRows::rows),
+            &self.shape,
+            n,
+        );
         super::apply_truncate_plan(&mut self.blocks, &plan);
         // NB: no `self.gpu.clear()` — the ring is the source of truth for a
         // ring-only decode tail; see the doc comment above.
