+++ crates/rmlx-kv-quant/src/storage/quant_v.rs
@@ -623,20 +623,7 @@ impl QuantV {
     /// [`super::clamp_truncate_target`] for why that is not defensive padding
     /// but the only correct reading for a ring-less store.
     pub fn truncate_to(&mut self, n: i32) {
-        let n = super::clamp_truncate_target(&self.shape, n);
-        let plan = super::truncate_plan(
-            self.blocks
-                .iter()
-                .map(|blk| super::block_rows(&blk.original_shape)),
-            &self.shape,
-            n,
-        );
-        super::apply_truncate_plan(&mut self.blocks, &plan);
-        // `get_mut` rather than `shape[2]`: the store shape is rank-4 by
-        // construction, and this is the bounds proof rather than a claim.
-        if let Some(seq) = self.shape.get_mut(2) {
-            *seq = n;
-        }
+        super::truncate_block_store(&mut self.blocks, &mut self.shape, n);
     }
 
     /// Dequantize all accumulated V slices to a flat f32 vec (CPU path) or
